reorganize router building to maybe ping redis less
